+++ crates/protocol/CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## 0.1.1-git
 
+### Minor
+
+- Add a `Connection` abstraction with `host` feature
+
 ### Patch
 
 - Update dependencies
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 1 -->
+<!-- Increment to skip CHANGELOG.md test: 3 -->